Rename calculations.ts to utils.ts
