+++ dashboard/15-final/app/ui/dashboard/revenue-chart.tsx
@@ -1,5 +1,5 @@
 import { Revenue } from '@/app/lib/definitions';
-import { generateYAxis } from '@/app/lib/calculations';
+import { generateYAxis } from '@/app/lib/utils';
 
 // This component is representational only.
 // For data visualization UI, check out:
